Merged PR 7602: Partitioned pool actor
This change allows the pool actor to be partitioned by an optional key. The pool actor's id/address can either be `fabric:/Audit.Firms/Services` or `fabric:/Audit.Engagements/Services?{firmguid}`. The servicetypeuri and actor id of the pool manager are now two different values in this PR as a result. There will now be a poolId that will include the entire address. The servicetypeuri will not include the query string param.